resolve DATA_DIRECTORY_DAEMON_HELP for masq UI
